Changed file names from CaptureViewCordShift folder and files to CaptureViewCoordShift and changed manifest file.
Created a test folder to work on CaptureViewCoordShift in.

Also adjusted input coordinates setup such that each input is shifted up by 1 so that coordinates given in image file names corespond to the actual coordinates in the mingrid model. For example input (0,0) would become (1,1) and correspond to the (0,0) location position in the image capture code.